check for subtree:"retain" AFTER call to view() to abort diffing
We want to abort diffing at a particular node but we can't receive that signal until after view() is called.
Currently the "retain" check occurs before the `{subtree:"retain"}` object is discovered via view().
It's unnecessary to let build() catch the "retain" on the next iteration when build() can return "cached" immediately upon discovery.

Shamefully I have not written tests for this, but I have anecdotally mixed and matched Components and Subcomponents, descendants and siblings, with different virtual elements marked "retain", using current-time markers. In all my test cases the right component(s) get diffed. A cached node that is only a string does not cause an error. And a component whose root node is a `{subtree:"retain"}` aborts the diff as well.

I did not test dynamic insertion of new components/subcomponents or their replacement. I'd expect inserting a new component or replacing one component for another that is the descendant of a subtree:retain might be a problem.  I presume `m.redraw.strategy('all')` would remedy that case. (or would it...?)

I hope I'm not missing the point, here.