Fix autoredraw on m.request SyntaxError
A m.request that results in a SyntaxError (or any built-in error which extends
from the Error object) will cause pendingRequests to become out of sync.
This is caused by m.deferred.onerror throwing an error which cause
m.endComputation() to never be called. However, looking at the tests, it
seems as if this was intentional, though not sure as to the reasoning
behind it.