Default to always using "beta" tag for rewrite
See https://docs.npmjs.com/files/package.json#publishconfig and
https://docs.npmjs.com/misc/config#tag

This way when `rewrite` starts being published users will need to
install it as `npm install mithril@beta` to get the rewrite. When people
run `npm install mithril` it'll use the default `latest` tag and that
will continue to point to the `0.2.x` branch until the rewrite is fully
ready.