Update README.md to clarify speed comparison chart
The use of the word "Performance" can be confusing in the speed comparison chart, because a shorter line means faster rendering and so higher performance -- compared to interpreting that performance line as operations per unit time at a first glance. "Rendering Time" as suggested on gitter by @brlewis makes clear that a shorter line is better.