added ability to specify data for URL-interpolation
Current implementation of `m.request` uses the data provided via `options.data` and replaces every found token which starts with a colon from that data. This commit adds a way to provide that data via `options.url.urldata`. To provide a backward compatible way the URL has to be provided as `options.url.rawurl`.

The idea is to separate request-body data from URL-interpolation data.