Clarify source of 2nd render pass re: preloading data
While the docs do say that a second render pass for preloaded data comes
from request completion, the example code for preloading data suggests that
a promise chain returned from `oninit` has a role to play in controlling the
second render pass.

The docs should make explicit where the redraw is initiated so the reader
does not mistakingly believe that `oninit()` retuning a promise changes
anything.