Split ospec, drop `mithril-stream` boilerplate
I'm keeping the source code for ospec for *now*, but this is deprecated
with a runtime warning on load. We've not documented it for a while, but
since I previously had people try the internal version, I need to be a
little more proactive in pushing people over to the separate npm
version. (It's actually up to date now.)