set window, document, requestAnimationFrame when missing
When importing mithril as a module in node, it blows up because there
is no window|document|requestAnimationFrame. This is particularly
problematic for ECMAScript modules (as opposed to CommmonJS Modules)
as there is no simple way to run code *before* importing a module. (There
are a number of hacks using the dynamic `import()` function, but these
aren't any fun.

This commit makes it such that when mithril appears to be imported from
inside node (by checking if `global.process.version` exists), it will
set the three variables to `undefined` if they don't exist.