Address some feedback around naming and (indirectly) `m.key`, optimize `m.use`
`m.use` is specially designed to be used a fair amount, and simulating it is relatively slow, so it's worth making a built-in.

`m.key` -> `m.keyed` provides for a simpler interface, one that you don't have to type `key` out explicitly for anymore. (Most frameworks require you to explicitly add a "key" attribute, while this approaches it more like a glorified `Map`.)