[refactor] Limit the condition of the option tag to "selected" attribute in isFormAttribute()
The condition of the option tag is intended to fix #1916 and should be evaluated in conjunction with the "selected" attribute.
This commit will improve code readability a bit, reduce the number of times to evaluate tags and also reduce the number of times to set the same attribute in option elements redundantly.