bundler: fix mangled strings caused by collision disambiguation in comment lines
This removes the mysterious suffixes in the comments and improves the readability of the bundled mithril.js. There is no impact on the code other than comments, as the processing is limited to lines containing only comments.
Trailing comments are not subject to this processing, but there are few trailing comments, and mangling does not seem to occur so far.