bundler: do not reset the suffix count if there is already a count
For some modules that are imported multiple times, it seems that the suffix count may be reset to 0. As a result, the suffix count does not increase, and a suffix may be added to the suffix, such as `mountRedraw00`. These double suffixes are annoying to fix, and it seems that `mountRedraw00` in the comments is corrected to `mountRedraw0`.
This commit prevents double suffixes by preventing the suffix count reset. Some suffixes in the bundled mithril.js will change, but this is because the suffixes are correctly incremented, and the code will not be broken (i.e., mithril.min.js will remain unchanged).