Process updates in reverse order of `old` instead of `vnodes`.
Because the DOM has not yet been arranged in `vnodes` order, `nextSibling` can point to the wrong DOM node. Since `insertBefore` uses `nextSibling`, nodes may be inserted in the wrong position when deletions and insertions occur. This commit fixes that by performing updates based on `old`, which reflects the actual DOM order.