Adjust the scanning range of `getNextSibling()`.
Previously, `getNextSibling()` scanned all the way to the end of the array. If the array contained vnodes that did not correspond to DOM nodes, or if newly inserted vnodes were present within the scan range, the scan could overrun and capture an incorrect nextSibling. This commit fixes that by limiting the scan range appropriately.