updateNodesFuzzer: Properly capture test data and comment out test cases that do not account for the heuristics in `updateNodes()`.
Previously, due to a bug that caused the same data to be reused, the fuzzer was not exercising the code adequately. It also included test cases that ignored the heuristics in `updateNodes()`, which had a very low probability of being triggered but, once they were, produced a large burst of false-positive test failures from that same data.
The commented-out sections are test cases for the number of append and insert operations related to the longest increasing subsequence algorithm, so it should be safe to leave them commented out for now.